Hoist tagging private packages out of publish package
