+++ packages/cli/src/commands/publish/__tests__/publishPackages.test.ts
@@ -38,11 +38,9 @@ describe("publishPackages", () => {
   describe("when isCI", () => {
     it("does not call out to npm to see if otp is required", async () => {
       await publishPackages({
-        cwd,
         packages: (await getPackages(cwd)).packages,
         access: "public",
-        preState: undefined,
-        tagPrivatePackages: true
+        preState: undefined
       });
       expect(npmUtils.getTokenIsRequired).not.toHaveBeenCalled();
     });
