+++ packages/cli/src/commands/publish/publishPackages.ts
@@ -1,7 +1,6 @@
 import { join } from "path";
 import semver from "semver";
 import chalk from "chalk";
-import * as git from "@changesets/git";
 import { AccessType } from "@changesets/types";
 import { Package } from "@manypkg/get-packages";
 import { info, warn } from "@changesets/logger";
@@ -19,7 +18,7 @@ type PkgInfo = {
   publishedVersions: string[];
 };
 
-type PublishedResult = {
+export type PublishedResult = {
   name: string;
   newVersion: string;
   published: boolean;
@@ -75,30 +74,20 @@ const getTwoFactorState = ({
 };
 
 export default async function publishPackages({
-  cwd,
   packages,
   access,
   otp,
   preState,
-  tag,
-  tagPrivatePackages
+  tag
 }: {
-  cwd: string;
   packages: Package[];
   access: AccessType;
   otp?: string;
   preState: PreState | undefined;
   tag?: string;
-  tagPrivatePackages: boolean;
-}): Promise<{
-  publishedPackages: PublishedResult[];
-  untaggedPrivatePackages: Omit<PublishedResult, "published">[];
-}> {
+}) {
   const packagesByName = new Map(packages.map((x) => [x.packageJson.name, x]));
   const publicPackages = packages.filter((pkg) => !pkg.packageJson.private);
-  const privatePackages = packages.filter(
-    pkg => pkg.packageJson.private && pkg.packageJson.version
-  );
   const unpublishedPackagesInfo = await getUnpublishedPackages(
     publicPackages,
     preState
@@ -115,7 +104,7 @@ export default async function publishPackages({
           isRequired: Promise.resolve(false)
         };
 
-  const npmPackagePublish = Promise.all(
+  return Promise.all(
     unpublishedPackagesInfo.map((pkgInfo) => {
       let pkg = packagesByName.get(pkgInfo.name)!;
       return publishAPackage(
@@ -126,58 +115,6 @@ export default async function publishPackages({
       );
     })
   );
-
-  const untaggedPrivatePackageReleases = tagPrivatePackages
-    ? getUntaggedPrivatePackages(privatePackages, cwd)
-    : Promise.resolve([]);
-
-  const result: {
-    publishedPackages: PublishedResult[];
-    untaggedPrivatePackages: PublishedResult[];
-  } = {
-    publishedPackages: await npmPackagePublish,
-    untaggedPrivatePackages: await untaggedPrivatePackageReleases
-  };
-
-  if (
-    result.publishedPackages.length === 0 &&
-    result.untaggedPrivatePackages.length === 0
-  ) {
-    warn("No unpublished projects to publish");
-  }
-
-  return result;
-}
-
-async function getUntaggedPrivatePackages(
-  privatePackages: Package[],
-  cwd: string
-) {
-  const packageWithTags = await Promise.all(
-    privatePackages.map(async privatePkg => {
-      const tagName = `${privatePkg.packageJson.name}@${privatePkg.packageJson.version}`;
-      const isMissingTag = !(
-        (await git.tagExists(tagName, cwd)) ||
-        (await git.remoteTagExists(tagName))
-      );
-
-      return { pkg: privatePkg, isMissingTag };
-    })
-  );
-
-  const untagged: PublishedResult[] = [];
-
-  for (const packageWithTag of packageWithTags) {
-    if (packageWithTag.isMissingTag) {
-      untagged.push({
-        name: packageWithTag.pkg.packageJson.name,
-        newVersion: packageWithTag.pkg.packageJson.version,
-        published: false
-      });
-    }
-  }
-
-  return untagged;
 }
 
 async function publishAPackage(
